Refactor -- Moved all area stuff to its own folder, and split it between more files.
