+++ Source/source/area/geometry.cpp
@@ -5,27 +5,23 @@
  * Pikmin is copyright (c) Nintendo.
  *
  * === FILE DESCRIPTION ===
- * Sector, edge, etc. classes, and related functions.
+ * Area geometry-related functions.
  */
 
 #define _USE_MATH_DEFINES
 
-#include <algorithm>
-#include <cfloat>
-#include <math.h>
-#include <queue>
-#include <unordered_set>
+#include <cmath>
+#include <vector>
 
-#include "sector.h"
+#include <allegro5/allegro_color.h>
 
-#include "functions.h"
-#include "game.h"
-#include "utils/geometry_utils.h"
-#include "utils/string_utils.h"
+#include "geometry.h"
 
+#include "../functions.h"
+#include "../utils/geometry_utils.h"
+#include "vertex.h"
 
-using std::unordered_set;
-using std::set;
+using std::vector;
 
 
 namespace GEOMETRY {
@@ -54,372 +50,6 @@ const float SMOOTHING_MAX_LENGTH = 100.0f;
 //Mobs can walk up sectors that are, at the most,
 //this high from the current one, as if climbing up steps.
 const float STEP_HEIGHT = 50;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Creates a blockmap.
- */
-blockmap::blockmap() :
-    n_cols(0),
-    n_rows(0) {
-    
-}
-
-
-/* ----------------------------------------------------------------------------
- * Clears the info of the blockmap.
- */
-void blockmap::clear() {
-    top_left_corner = point();
-    edges.clear();
-    sectors.clear();
-    n_cols = 0;
-    n_rows = 0;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the block column in which an X coordinate is contained.
- * Returns INVALID on error.
- * x:
- *   X coordinate.
- */
-size_t blockmap::get_col(const float x) const {
-    if(x < top_left_corner.x) return INVALID;
-    float final_x = (x - top_left_corner.x) / GEOMETRY::BLOCKMAP_BLOCK_SIZE;
-    if(final_x >= n_cols) return INVALID;
-    return final_x;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Obtains a list of edges that are within the specified rectangular region.
- * Returns true on success, false on error.
- * tl:
- *   Top-left coordinates of the region.
- * br:
- *   Bottom-right coordinates of the region.
- * edges:
- *   Set to fill the edges into.
- */
-bool blockmap::get_edges_in_region(
-    const point &tl, const point &br, set<edge*> &edges
-) const {
-
-    size_t bx1 = game.cur_area_data.bmap.get_col(tl.x);
-    size_t bx2 = game.cur_area_data.bmap.get_col(br.x);
-    size_t by1 = game.cur_area_data.bmap.get_row(tl.y);
-    size_t by2 = game.cur_area_data.bmap.get_row(br.y);
-    
-    if(
-        bx1 == INVALID || bx2 == INVALID ||
-        by1 == INVALID || by2 == INVALID
-    ) {
-        //Out of bounds.
-        return false;
-    }
-    
-    for(size_t bx = bx1; bx <= bx2; ++bx) {
-        for(size_t by = by1; by <= by2; ++by) {
-        
-            vector<edge*> &block_edges =
-                game.cur_area_data.bmap.edges[bx][by];
-                
-            for(size_t e = 0; e < block_edges.size(); ++e) {
-                edges.insert(block_edges[e]);
-            }
-        }
-    }
-    
-    return true;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the block row in which a Y coordinate is contained.
- * Returns INVALID on error.
- * y:
- *   Y coordinate.
- */
-size_t blockmap::get_row(const float y) const {
-    if(y < top_left_corner.y) return INVALID;
-    float final_y = (y - top_left_corner.y) / GEOMETRY::BLOCKMAP_BLOCK_SIZE;
-    if(final_y >= n_rows) return INVALID;
-    return final_y;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the top-left coordinates for the specified column and row.
- * col:
- *   Column to check.
- * row:
- *   Row to check.
- */
-point blockmap::get_top_left_corner(const size_t col, const size_t row) const {
-    return
-        point(
-            col * GEOMETRY::BLOCKMAP_BLOCK_SIZE + top_left_corner.x,
-            row * GEOMETRY::BLOCKMAP_BLOCK_SIZE + top_left_corner.y
-        );
-}
-
-
-/* ----------------------------------------------------------------------------
- * Creates an edge.
- * v1:
- *   Its first vertex.
- * v2:
- *   Its second vertex.
- */
-edge::edge(size_t v1, size_t v2) :
-    wall_shadow_length(LARGE_FLOAT),
-    wall_shadow_color(GEOMETRY::SHADOW_DEF_COLOR),
-    ledge_smoothing_length(0),
-    ledge_smoothing_color(GEOMETRY::SMOOTHING_DEF_COLOR) {
-    
-    vertexes[0] = vertexes[1] = NULL;
-    sectors[0] = sectors[1] = NULL;
-    sector_nrs[0] = sector_nrs[1] = INVALID;
-    
-    vertex_nrs[0] = v1; vertex_nrs[1] = v2;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Clones an edge's properties onto another,
- * not counting the sectors or vertexes.
- * destination:
- *   Edge to clone the data into.
- */
-void edge::clone(edge* destination) const {
-    destination->wall_shadow_length = wall_shadow_length;
-    destination->wall_shadow_color = wall_shadow_color;
-    destination->ledge_smoothing_length = ledge_smoothing_length;
-    destination->ledge_smoothing_color = ledge_smoothing_color;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the sector that ISN'T the specified one.
- * s_ptr:
- *   The sector that ISN'T the one to return.
- */
-sector* edge::get_other_sector(const sector* s_ptr) const {
-    if(sectors[0] == s_ptr) return sectors[1];
-    return sectors[0];
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the vertex that ISN'T the specified one.
- * v_ptr:
- *   The vertex that ISN'T the one to return.
- */
-vertex* edge::get_other_vertex(const vertex* v_ptr) const {
-    if(vertexes[0] == v_ptr) return vertexes[1];
-    return vertexes[0];
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns which side has the specified sector, or INVALID if neither.
- * s_ptr:
- *   Sector to check.
- */
-size_t edge::get_side_with_sector(sector* s_ptr) const {
-    for(unsigned char s = 0; s < 2; ++s) {
-        if(sectors[s] == s_ptr) return s;
-    }
-    return INVALID;
-}
-
-
-/* ----------------------------------------------------------------------------
- * If the specified edge and this one are not neighbors, returns NULL.
- * Otherwise, returns the vertex that binds them.
- * other:
- *   Edge to check as a neighbor.
- */
-vertex* edge::has_neighbor(edge* other) const {
-    for(size_t v1 = 0; v1 < 2; ++v1) {
-        for(size_t v2 = 0; v2 < 2; ++v2) {
-            if(vertexes[v1] == other->vertexes[v2]) {
-                return vertexes[v1];
-            }
-        }
-    }
-    return NULL;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether or not an edge is valid.
- * An edge is valid if it has non-NULL vertexes.
- */
-bool edge::is_valid() const {
-    if(!vertexes[0]) return false;
-    if(!vertexes[1]) return false;
-    return true;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Removes the edge from its sectors, but doesn't mark
- * the sectors as "none".
- * Returns the edge number.
- */
-size_t edge::remove_from_sectors() {
-    size_t e_nr = INVALID;
-    for(unsigned char s = 0; s < 2; ++s) {
-        sector* s_ptr = sectors[s];
-        if(!s_ptr) continue;
-        for(size_t e = 0; e < s_ptr->edges.size(); ++e) {
-            edge* e_ptr = s_ptr->edges[e];
-            if(e_ptr == this) {
-                s_ptr->edges.erase(s_ptr->edges.begin() + e);
-                auto nr_it = s_ptr->edge_nrs.begin() + e;
-                e_nr = *nr_it;
-                s_ptr->edge_nrs.erase(nr_it);
-                break;
-            }
-        }
-        sectors[s] = NULL;
-        sector_nrs[s] = INVALID;
-    }
-    return e_nr;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Removes the edge from its vertexes, but doesn't mark
- * the vertexes as "none".
- * Returns the edge number.
- */
-size_t edge::remove_from_vertexes() {
-    size_t e_nr = INVALID;
-    for(unsigned char v = 0; v < 2; ++v) {
-        vertex* v_ptr = vertexes[v];
-        if(!v_ptr) continue;
-        for(size_t e = 0; e < v_ptr->edges.size(); ++e) {
-            edge* e_ptr = v_ptr->edges[e];
-            if(e_ptr == this) {
-                v_ptr->edges.erase(v_ptr->edges.begin() + e);
-                auto nr_it = v_ptr->edge_nrs.begin() + e;
-                e_nr = *nr_it;
-                v_ptr->edge_nrs.erase(nr_it);
-                break;
-            }
-        }
-        vertexes[v] = NULL;
-        vertex_nrs[v] = INVALID;
-    }
-    return e_nr;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Swaps the two vertexes of the edge around. It also swaps the sectors,
- * so that they still point in the right direction.
- */
-void edge::swap_vertexes() {
-    std::swap(vertexes[0], vertexes[1]);
-    std::swap(vertex_nrs[0], vertex_nrs[1]);
-    std::swap(sectors[0], sectors[1]);
-    std::swap(sector_nrs[0], sector_nrs[1]);
-}
-
-
-/* ----------------------------------------------------------------------------
- * Transfers this edge from one sector to a different one.
- * from:
- *   Sector to transfer from.
- * to:
- *   Sector to transfer to.
- * to_nr:
- *   Number of the sector to transfer to.
- * edge_nr:
- *   Number of the current edge.
- */
-void edge::transfer_sector(
-    sector* from, sector* to, const size_t to_nr, const size_t edge_nr
-) {
-    size_t index = get_side_with_sector(from);
-    engine_assert(
-        index != INVALID,
-        i2s(to_nr)
-    );
-    
-    sectors[index] = to;
-    sector_nrs[index] = to_nr;
-    
-    if(from) from->remove_edge(this);
-    if(to) to->add_edge(this, edge_nr);
-}
-
-
-/* ----------------------------------------------------------------------------
- * Creates an edge intersection info structure.
- * e1:
- *   First edge in the intersection.
- * e2:
- *   Second edge in the intersection.
- */
-edge_intersection::edge_intersection(edge* e1, edge* e2) :
-    e1(e1),
-    e2(e2) {
-    
-}
-
-
-/* ----------------------------------------------------------------------------
- * Checks whether the edge intersection contains the specified edge.
- * e:
- *   Edge to check.
- */
-bool edge_intersection::contains(edge* e) {
-    return e1 == e || e2 == e;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Creates a mob generation structure.
- * pos:
- *   Coordinates.
- * type:
- *   The mob type.
- * angle:
- *   Angle it is facing.
- * vars:
- *   String representation of the script vars.
- */
-mob_gen::mob_gen(
-    const point &pos, mob_type* type, const float angle, const string &vars
-) :
-    type(type),
-    pos(pos),
-    angle(angle),
-    vars(vars),
-    stored_inside(INVALID) {
-    
-}
-
-
-/* ----------------------------------------------------------------------------
- * Clones the properties of this mob generator onto another mob generator.
- * destination:
- *   Mob generator to clone the data into.
- * position:
- *   If true, the position is included too.
- */
-void mob_gen::clone(mob_gen* destination, const bool include_position) const {
-    destination->angle = angle;
-    if(include_position) destination->pos = pos;
-    destination->type = type;
-    destination->vars = vars;
-    destination->link_nrs = link_nrs;
-    destination->stored_inside = stored_inside;
 }
 
 
@@ -734,347 +364,6 @@ vertex* polygon::get_rightmost_vertex() const {
 
 
 /* ----------------------------------------------------------------------------
- * Creates a sector.
- */
-sector::sector() :
-    type(SECTOR_TYPE_NORMAL),
-    is_bottomless_pit(false),
-    z(0),
-    brightness(GEOMETRY::DEF_SECTOR_BRIGHTNESS),
-    fade(false),
-    hazard_floor(true),
-    liquid_drain_left(0),
-    draining_liquid(false) {
-    
-}
-
-
-/* ----------------------------------------------------------------------------
- * Destroys a sector.
- */
-sector::~sector() {
-    for(size_t t = 0; t < 2; ++t) {
-        if(texture_info.bitmap && texture_info.bitmap != game.bmp_error) {
-            game.bitmaps.detach(texture_info.file_name);
-        }
-    }
-}
-
-
-/* ----------------------------------------------------------------------------
- * Adds an edge to the sector's list of edges, if it's not there already.
- * e_ptr:
- *   Edge to add.
- * e_nr:
- *   Index number of the edge to add.
- */
-void sector::add_edge(edge* e_ptr, const size_t e_nr) {
-    for(size_t i = 0; i < edges.size(); ++i) {
-        if(edges[i] == e_ptr) {
-            return;
-        }
-    }
-    edges.push_back(e_ptr);
-    edge_nrs.push_back(e_nr);
-}
-
-
-/* ----------------------------------------------------------------------------
- * Calculates the bounding box coordinates and saves them
- * in the object's bbox variables.
- */
-void sector::calculate_bounding_box() {
-    if(edges.empty()) {
-        //Unused sector... This shouldn't exist.
-        bbox[0] = point();
-        bbox[1] = point();
-        return;
-    }
-    
-    bbox[0].x = edges[0]->vertexes[0]->x;
-    bbox[1].x = bbox[0].x;
-    bbox[0].y = edges[0]->vertexes[0]->y;
-    bbox[1].y = bbox[0].y;
-    
-    for(size_t e = 0; e < edges.size(); ++e) {
-        for(unsigned char v = 0; v < 2; ++v) {
-            point coords(
-                edges[e]->vertexes[v]->x,
-                edges[e]->vertexes[v]->y
-            );
-            
-            bbox[0].x = std::min(bbox[0].x, coords.x);
-            bbox[1].x = std::max(bbox[1].x, coords.x);
-            bbox[0].y = std::min(bbox[0].y, coords.y);
-            bbox[1].y = std::max(bbox[1].y, coords.y);
-        }
-    }
-}
-
-
-/* ----------------------------------------------------------------------------
- * Clones a sector's properties onto another,
- * not counting the list of edges, bounding box, or bitmap
- * (the file name is cloned too, though).
- * destination:
- *   Sector to clone the data into.
- */
-void sector::clone(sector* destination) const {
-    destination->type = type;
-    destination->is_bottomless_pit = is_bottomless_pit;
-    destination->z = z;
-    destination->tag = tag;
-    destination->hazards = hazards;
-    destination->hazard_floor = hazard_floor;
-    destination->hazards_str = hazards_str;
-    destination->brightness = brightness;
-    destination->texture_info.scale = texture_info.scale;
-    destination->texture_info.translation = texture_info.translation;
-    destination->texture_info.rot = texture_info.rot;
-    destination->texture_info.tint = texture_info.tint;
-    destination->fade = fade;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Fills a vector with neighboring sectors, recursively, but only if they
- * meet certain criteria.
- * condition:
- *   Function that accepts a sector and checks its criteria. This
- *   function must return true if accepted, false if not.
- * sector_list:
- *   List of sectors to be filled. Also doubles as the list of visited sectors.
- */
-void sector::get_neighbor_sectors_conditionally(
-    const std::function<bool(sector* s_ptr)> &condition,
-    vector<sector*> &sector_list
-) {
-
-    //If this sector is already on the list, skip.
-    for(size_t s = 0; s < sector_list.size(); ++s) {
-        if(sector_list[s] == this) return;
-    }
-    
-    //If this sector is not eligible, return.
-    if(!condition(this)) return;
-    
-    //This sector is valid!
-    sector_list.push_back(this);
-    
-    //Now check its neighbors.
-    edge* e_ptr = NULL;
-    sector* other_s = NULL;
-    for(size_t e = 0; e < edges.size(); ++e) {
-        e_ptr = edges[e];
-        other_s = e_ptr->get_other_sector(this);
-        if(!other_s) continue;
-        
-        other_s->get_neighbor_sectors_conditionally(condition, sector_list);
-    }
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the vertex farthest to the right in a sector.
- */
-vertex* sector::get_rightmost_vertex() const {
-    vertex* rightmost = NULL;
-    
-    for(size_t e = 0; e < edges.size(); ++e) {
-        edge* e_ptr = edges[e];
-        if(!rightmost) rightmost = e_ptr->vertexes[0];
-        else {
-            rightmost = ::get_rightmost_vertex(e_ptr->vertexes[0], rightmost);
-            rightmost = ::get_rightmost_vertex(e_ptr->vertexes[1], rightmost);
-        }
-    }
-    
-    return rightmost;
-}
-
-
-/* ----------------------------------------------------------------------------
- * If texture merging is required, this returns what two neighboring sectors
- * will be used for it.
- * s1:
- *   Receives the first sector.
- * s2:
- *   Receives the second sector.
- */
-void sector::get_texture_merge_sectors(sector** s1, sector** s2) const {
-    //Check all edges to find which two textures need merging.
-    edge* e_ptr = NULL;
-    sector* neighbor = NULL;
-    map<sector*, dist> neighbors;
-    sector* texture_sector[2] = {NULL, NULL};
-    
-    //The two neighboring sectors with the lenghtiest edges are picked.
-    //So save all sector/length pairs.
-    //Sectors with different heights from the current one are also saved,
-    //but they have lower priority compared to same-heigh sectors.
-    for(size_t e = 0; e < edges.size(); ++e) {
-        e_ptr = edges[e];
-        bool valid = true;
-        
-        neighbor = e_ptr->get_other_sector(this);
-        
-        if(neighbor) {
-            if(neighbor->fade) valid = false;
-        }
-        
-        if(valid) {
-            neighbors[neighbor] +=
-                dist(
-                    point(e_ptr->vertexes[0]->x, e_ptr->vertexes[0]->y),
-                    point(e_ptr->vertexes[1]->x, e_ptr->vertexes[1]->y)
-                );
-        }
-    }
-    
-    //Find the two lengthiest ones.
-    vector<std::pair<dist, sector*> > neighbors_vec;
-    for(auto &n : neighbors) {
-        neighbors_vec.push_back(
-            std::make_pair(
-                //Yes, we do need these casts, for g++.
-                (dist) (n.second), (sector*) (n.first)
-            )
-        );
-    }
-    sort(
-        neighbors_vec.begin(), neighbors_vec.end(),
-    [this] (std::pair<dist, sector*> p1, std::pair<dist, sector*> p2) -> bool {
-        return p1.first < p2.first;
-    }
-    );
-    if(neighbors_vec.size() >= 1) {
-        texture_sector[0] = neighbors_vec.back().second;
-    }
-    if(neighbors_vec.size() >= 2) {
-        texture_sector[1] = neighbors_vec[neighbors_vec.size() - 2].second;
-    }
-    
-    if(!texture_sector[1] && texture_sector[0]) {
-        //0 is always the bottom one. If we're fading into nothingness,
-        //we should swap first.
-        std::swap(texture_sector[0], texture_sector[1]);
-    } else if(!texture_sector[1]) {
-        //Nothing to draw.
-        return;
-    } else if(texture_sector[1]->is_bottomless_pit) {
-        std::swap(texture_sector[0], texture_sector[1]);
-    }
-    
-    *s1 = texture_sector[0];
-    *s2 = texture_sector[1];
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether a sector's vertexes are ordered clockwise or not.
- */
-bool sector::is_clockwise() const {
-    vector<vertex*> vertexes;
-    for(size_t e = 0; e < edges.size(); ++e) {
-        vertexes.push_back(edges[e]->vertexes[0]);
-    }
-    return is_polygon_clockwise(vertexes);
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether a point is inside a sector by checking its triangles.
- * p:
- *   Point to check.
- */
-bool sector::is_point_in_sector(const point &p) const {
-    for(size_t t = 0; t < triangles.size(); ++t) {
-        const triangle* t_ptr = &triangles[t];
-        if(
-            is_point_in_triangle(
-                p,
-                point(t_ptr->points[0]->x, t_ptr->points[0]->y),
-                point(t_ptr->points[1]->x, t_ptr->points[1]->y),
-                point(t_ptr->points[2]->x, t_ptr->points[2]->y),
-                false
-            )
-        ) {
-            return true;
-        }
-    }
-    
-    return false;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Removes an edge from a sector's list of edges, if it is there.
- * e_ptr:
- *   Edge to remove.
- */
-void sector::remove_edge(edge* e_ptr) {
-    size_t i = 0;
-    for(; i < edges.size(); ++i) {
-        if(edges[i] == e_ptr) {
-            edges.erase(edges.begin() + i);
-            edge_nrs.erase(edge_nrs.begin() + i);
-            return;
-        }
-    }
-}
-
-
-/* ----------------------------------------------------------------------------
- * Creates a sector texture's info struct.
- */
-sector_texture_info::sector_texture_info() :
-    scale(1.0, 1.0),
-    rot(0),
-    bitmap(nullptr),
-    tint(COLOR_WHITE) {
-}
-
-
-
-/* ----------------------------------------------------------------------------
- * Creates a tree shadow.
- * center:
- *   Center coordinates.
- * size:
- *   Width and height.
- * angle:
- *   Angle it is rotated by.
- * alpha:
- *   How opaque it is [0-255].
- * file_name:
- *   Name of the file with the tree shadow's texture.
- * sway:
- *   Multiply the sway distance by this much, horizontally and vertically.
- */
-tree_shadow::tree_shadow(
-    const point &center, const point &size, const float angle,
-    const unsigned char alpha, const string &file_name, const point &sway
-) :
-    file_name(file_name),
-    bitmap(nullptr),
-    center(center),
-    size(size),
-    angle(angle),
-    alpha(alpha),
-    sway(sway) {
-    
-}
-
-
-/* ----------------------------------------------------------------------------
- * Destroys a tree shadow.
- */
-tree_shadow::~tree_shadow() {
-    game.textures.detach(file_name);
-}
-
-
-/* ----------------------------------------------------------------------------
  * Creates a triangle.
  * v1:
  *   First vertex.
@@ -1087,158 +376,6 @@ triangle::triangle(vertex* v1, vertex* v2, vertex* v3) {
     points[0] = v1;
     points[1] = v2;
     points[2] = v3;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Creates a vertex.
- * x:
- *   X coordinate.
- * y:
- *   Y coordinate.
- */
-vertex::vertex(float x, float y) :
-    x(x),
-    y(y) {
-    
-}
-
-
-/* ----------------------------------------------------------------------------
- * Adds an edge to the vertex's list of edges, if it's not there already.
- * e_ptr:
- *   Edge to add.
- * e_nr:
- *   Index number of the edge to add.
- */
-void vertex::add_edge(edge* e_ptr, const size_t e_nr) {
-    for(size_t i = 0; i < edges.size(); ++i) {
-        if(edges[i] == e_ptr) {
-            return;
-        }
-    }
-    edges.push_back(e_ptr);
-    edge_nrs.push_back(e_nr);
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns the edge that has the specified vertex as a neighbor of this vertex.
- * Returns NULL if not found.
- * neighbor:
- *   The neighbor vertex to check.
- */
-edge* vertex::get_edge_by_neighbor(vertex* neighbor) const {
-    for(size_t e = 0; e < edges.size(); ++e) {
-        if(edges[e]->get_other_vertex(this) == neighbor) {
-            return edges[e];
-        }
-    }
-    return NULL;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether or not this vertex has the specified edge in its list.
- * e_ptr:
- *   Edge to check.
- */
-bool vertex::has_edge(edge* e_ptr) const {
-    for(size_t e = 0; e < edges.size(); ++e) {
-        if(edges[e] == e_ptr) return true;
-    }
-    return false;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether or not this vertex is a second-degree neighbor to the
- * specified vertex. i.e. they have a shared neighbor between them.
- * other_v:
- *   The vertex to compare against.
- * first_neighbor:
- *   Return the common neighbor between them here, if the result is true.
- */
-bool vertex::is_2nd_degree_neighbor(
-    vertex* other_v, vertex** first_neighbor
-) const {
-    //Let's crawl forward through all edges and stop at the second level.
-    //If other_v is at that distance, then we found it!
-    
-    for(size_t e1 = 0; e1 < edges.size(); ++e1) {
-        vertex* next_v = edges[e1]->get_other_vertex(this);
-        
-        for(size_t e2 = 0; e2 < next_v->edges.size(); ++e2) {
-            if(next_v->edges[e2]->get_other_vertex(next_v) == other_v) {
-                *first_neighbor = next_v;
-                return true;
-            }
-        }
-        
-    }
-    return false;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether or not this vertex is a second-degree neighbor to the
- * specified edge. i.e. one of the vertex's neighbors is used by the edge.
- * other_e:
- *   The edge to compare against.
- * first_neighbor:
- *   Return the common neighbor between them here, if the result is true.
- */
-bool vertex::is_2nd_degree_neighbor(
-    edge* other_e, vertex** first_neighbor
-) const {
-    //Let's crawl forward through all edges and stop at the second level.
-    //If other_e is at that distance, then we found it!
-    
-    for(size_t e1 = 0; e1 < edges.size(); ++e1) {
-        vertex* next_v = edges[e1]->get_other_vertex(this);
-        
-        for(size_t e2 = 0; e2 < next_v->edges.size(); ++e2) {
-            if(next_v->edges[e2] == other_e) {
-                *first_neighbor = next_v;
-                return true;
-            }
-        }
-        
-    }
-    return false;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns whether or not this vertex is a neighbor to the
- * specified vertex. i.e. they have a shared edge between them.
- * other_v:
- *   The vertex to compare against.
- */
-bool vertex::is_neighbor(vertex* other_v) const {
-    for(size_t e = 0; e < edges.size(); ++e) {
-        if(edges[e]->get_other_vertex(this) == other_v) {
-            return true;
-        }
-    }
-    return false;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Removes an edge from a vertex's list of edges, if it is there.
- * e_ptr:
- *   Edge to remove.
- */
-void vertex::remove_edge(edge* e_ptr) {
-    size_t i = 0;
-    for(; i < edges.size(); ++i) {
-        if(edges[i] == e_ptr) {
-            edges.erase(edges.begin() + i);
-            edge_nrs.erase(edge_nrs.begin() + i);
-            return;
-        }
-    }
 }
 
 
@@ -1529,71 +666,6 @@ vertex* get_rightmost_vertex(vertex* v1, vertex* v2) {
     if(v1->x > v2->x) return v1;
     if(v1->x == v2->x && v1->y < v2->y) return v1;
     return v2;
-}
-
-
-/* ----------------------------------------------------------------------------
- * Returns which sector the specified point belongs to.
- * p:
- *   Coordinates of the point.
- * sector_nr:
- *   If not NULL, the number of the sector on the area map is placed here.
- *   The number will not be set if the search is using the blockmap.
- * use_blockmap:
- *   If true, use the blockmap to search.
- *   This provides faster results, but the blockmap must be built.
- */
-sector* get_sector(
-    const point &p, size_t* sector_nr, const bool use_blockmap
-) {
-
-    if(use_blockmap) {
-    
-        size_t col = game.cur_area_data.bmap.get_col(p.x);
-        size_t row = game.cur_area_data.bmap.get_row(p.y);
-        if(col == INVALID || row == INVALID) return NULL;
-        
-        unordered_set<sector*>* sectors =
-            &game.cur_area_data.bmap.sectors[col][row];
-            
-        if(sectors->size() == 1) return *sectors->begin();
-        
-        for(auto s : (*sectors)) {
-        
-            if(!s) {
-                continue;
-            }
-            if(s->is_point_in_sector(p)) {
-                return s;
-            }
-        }
-        
-        return NULL;
-        
-    } else {
-    
-        for(size_t s = 0; s < game.cur_area_data.sectors.size(); ++s) {
-            sector* s_ptr = game.cur_area_data.sectors[s];
-            
-            if(
-                p.x < s_ptr->bbox[0].x ||
-                p.x > s_ptr->bbox[1].x ||
-                p.y < s_ptr->bbox[0].y ||
-                p.y > s_ptr->bbox[1].y
-            ) {
-                continue;
-            }
-            if(s_ptr->is_point_in_sector(p)) {
-                if(sector_nr) *sector_nr = s;
-                return s_ptr;
-            }
-            
-        }
-        
-        if(sector_nr) *sector_nr = INVALID;
-        return NULL;
-        
-    }
 }
